Carry a kernel interface index at its real width
An OS interface index is a 32-bit unsigned number that, on Linux, comes
from a counter which only ever increases - a long-lived host that churns
veth or container interfaces will issue values well past 65535. PMIx
carried one in a uint16_t field and handed it back through helpers that
did not agree on a width:

  pmix_pif_t.if_kernel_index      uint16_t
  pmix_ifnametokindex()           int16_t
  pmix_ifaddrtokindex()           int16_t
  pmix_ifindextokindex()          int
  pmix_ifkindextoname()           int parameter
  pmix_ifkindextoaddr()           int parameter
  pmix_ifmatches()                int parameter

The disagreement bites well before the field overflows. An index of
40000 stores in the uint16_t perfectly well, and pmix_ifindextokindex()
returns it correctly - but pmix_ifnametokindex() returns it as -25536,
and every caller of these helpers screens the result with "if (0 >
kindex)". So the same interface was simultaneously findable by list
index and missing by name, which is a silent wrong answer rather than an
error: an if_include entry naming that interface is quietly skipped.

Everything now uses one width. The field is a uint32_t, both int16_t
returns become int - which is what every consumer of a kernel index
already took, and what PRRTE's own kindex variables already are - and
the constructor's unassigned sentinel becomes UINT32_MAX.

Widening the field made the compiler point out four comparisons that
were relying on the old integer promotion, and they are worth more than
a cast: pmix_ifkindextoname(), pmix_ifkindextoaddr() and pmix_ifmatches()
all take a signed parameter that callers legitimately feed a -1 from a
failed lookup, and converting that to unsigned would have produced
UINT32_MAX - the exact value the constructor leaves in an entry no
component has filled in. Each now rejects a negative index up front.

test/unit/util/util_if.c builds its interface list by hand, which is the
only way to exercise this: no developer's machine has an interface index
above 32767. Both halves were verified by re-breaking - restoring the
int16_t return fails three of the new cases (nametokindex disagreeing
with indextokindex at 40000), and restoring the uint16_t field fails the
70000 case on its own.

Note for anything building against these headers: pmix_pif_t is in the
installed src/util/pmix_if.h and PRRTE's oob does walk pmix_if_list and
construct pmix_pif_t objects directly, so it needs a rebuild against
this. It needs no source change - its three uses of if_kernel_index all
go through an int already - and it calls neither of the two functions
whose return type changed.

+++ src/mca/pif/bsdx_ipv4/pif_bsdx.c
@@ -161,7 +161,7 @@ static int if_bsdx_open(void)
         }
         intf->if_flags = cur_ifaddrs->ifa_flags;
 
-        intf->if_kernel_index = (uint16_t) if_nametoindex(cur_ifaddrs->ifa_name);
+        intf->if_kernel_index = if_nametoindex(cur_ifaddrs->ifa_name);
 
         pmix_list_append(&pmix_if_list, &(intf->super));
     } /*  of for loop over ifaddrs list */
